src/Twitter_simple_download_and_preprocess/requirements.txt
pandas
nltk
joblib
regex
scikit-learn
tqdm
mlflow
python-dotenv
